payment-request/payment-request-show-accept-promise-returned.html

<!DOCTYPE HTML>
<meta charset='utf-8'>
<title>PaymentRequest: when calling show() the acceptPromise should get created and returned.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<p></p>

<script>
async_test((t) => {
  const paymentArgs = [[{supportedMethods: ['non-supported-payment-method']}], {total: {label: 'label', amount: {currency: 'USD', value: '5.00'}}}];

  onload = t.step_func_done(() => {
    let pr = new PaymentRequest(paymentArgs)
    let p = pr.show()
    assert_true(p instanceof Promise);
  });
});
